feat/print-button
print input and translated text in new window
