+++ src/lib/Translator.svelte
@@ -1,47 +1,14 @@
 <script lang="ts">
+    import PrintButton from "./PrintButton.svelte";
+
     let translatedText = $state("");
     let inputText = $state("");
-
-    const printContent = () => {
-        console.log(translatedText)
-        try {
-            const newValue = translatedText;
-            const newWindow = window.open('', '_blank');
-
-            if (newWindow) {
-                newWindow.document.write(`
-          <!DOCTYPE html>
-          <html lang="en">
-            <head>
-              <meta charset="UTF-8">
-              <meta name="viewport" content="width=device-width, initial-scale=1.0">
-              <title>Printable Content</title>
-              <style>
-                @font-face { font-family: "Dovahkul"; src: url("assets/fonts/dragon_alphabet.woff") format('woff'); }
-                body { font-family: "Dovahkul", sans-serif; line-height: 1.6; padding: 20px; }
-              </style>
-            </head>
-            <body>
-              ${newValue}
-            </body>
-          </html>`);
-                newWindow.document.close();
-                newWindow.print();
-                newWindow.close();
-            } else {
-                console.error('Failed to open new window');
-            }
-        } catch (error) {
-            console.error('An error occurred:', error);
-        }
-    };
 
     $effect(() => {
         translatedText = inputText
     });
 
     /* TODO: replace words in real time using a function*/
-    /* TODO: function to export or print text */
     /*TODO: english <-> Dovahzul toggle switch class in .translated-text and function used to translate will need to reverse */
 </script>
 
@@ -49,9 +16,7 @@
 <section>
     <textarea class="translated-text">{translatedText}</textarea>
     <input bind:value={inputText} placeholder="Enter test to translate here..." class="text-input"/>
-    <button type="submit" class="print-button" onclick={printContent}>
-        Print translated content
-    </button>
+    <PrintButton inputText={inputText} translatedText={translatedText} />
 </section>
 
 <style>
@@ -78,9 +43,5 @@
         height: 10vh;
         text-align: center;
         font-size: 1.5rem;
-    }
-
-    .print-button {
-        margin: 5vh;
     }
 </style>
